Give up for now
This is harder than it should be. It's not documented how to save an
arbitrary file to the local system. I'm bored with this, there are
more interesting challenges. I'll pick this up again later.

There are still some nice code amendments and useful links so I'm
merging this rather than leaving it alone.